introduced namespaces & common superclass for BEs
