+++ CMakeLists.txt
@@ -54,6 +54,9 @@ endif()
 if(THORIN_PROFILE)
     set(THORIN_ENABLE_PROFILING TRUE)
 endif()
+if(LLVM_FOUND)
+    set(THORIN_ENABLE_LLVM TRUE)
+endif()
 if(RV_FOUND)
     set(THORIN_ENABLE_RV TRUE)
 endif()
