+++ src/thorin/be/llvm/cpu.h
@@ -3,7 +3,7 @@
 
 #include "thorin/be/llvm/llvm.h"
 
-namespace thorin {
+namespace thorin::llvm_be {
 
 class CPUCodeGen : public CodeGen {
 public:
