+++ src/thorin/config.h.in
@@ -3,6 +3,7 @@
 
 #cmakedefine01 THORIN_ENABLE_CHECKS
 #cmakedefine01 THORIN_ENABLE_PROFILING
+#cmakedefine01 THORIN_ENABLE_LLVM
 #cmakedefine01 THORIN_ENABLE_RV
 
 #endif
